admin/requirements.txt
django
django-environ
psycopg2-binary
django-ordered-model